requirements.txt
Django==2.2.16
django-cors-headers==3.5.0
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
